update: tidy up code
