+++ app/src/main/java/com/google/android/gms/example/appopenexample/callback/OnShowAdCompleteListener.kt
@@ -1,0 +1,20 @@
+package com.google.android.gms.example.appopenexample.callback
+
+/**
+ * Created by Faisal Amir on 24/10/22
+ * -----------------------------------------
+ * E-mail   : faisalamircs@gmail.com
+ * Github   : github.com/amirisback
+ * -----------------------------------------
+ * Copyright (C) Frogobox ID / amirisback
+ * All rights reserved
+ */
+
+
+/**
+ * Interface definition for a callback to be invoked when an app open ad is complete (i.e.
+ * dismissed or fails to show).
+ */
+interface OnShowAdCompleteListener {
+    fun onShowAdComplete()
+}
