+++ app/src/main/java/com/google/android/gms/example/appopenexample/util/Ext.kt
@@ -1,0 +1,24 @@
+package com.google.android.gms.example.appopenexample.util
+
+import android.app.Activity
+import android.content.Context
+import android.widget.Toast
+
+/**
+ * Created by Faisal Amir on 24/10/22
+ * -----------------------------------------
+ * E-mail   : faisalamircs@gmail.com
+ * Github   : github.com/amirisback
+ * -----------------------------------------
+ * Copyright (C) Frogobox ID / amirisback
+ * All rights reserved
+ */
+
+
+fun Activity.showToast(message: String) {
+    Toast.makeText(this, message, Toast.LENGTH_SHORT).show()
+}
+
+fun Context.showToast(message: String) {
+    Toast.makeText(this, message, Toast.LENGTH_SHORT).show()
+}
